% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all.R
\name{download_all}
\alias{download_all}
\title{Download all the links at once}
\usage{
download_all(DF)
}
\description{
To download all the links in the dataframe created with meetings().
First, it checks if the file do not exist, than save it. If the file is already
available, it jumps to the next line and proceed it again.
}
\examples{
download_all( DF )

}
