% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata from text column from dataframe}
\usage{
extract_metadata(DF)
}
\arguments{
\item{DF}{the dataframe generated by from_html_2_df}
}
\description{
input is DF generated by from_html_2_df(), export a DF with more columns
}
\examples{
file_html <- list.files("inst/extdata/", full.names=TRUE)
NT_html <- xml2::read_html(file_html,encoding = "utf8")
extract_metadata(NT_html)
}
