% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{from_html_2_df}
\alias{from_html_2_df}
\title{from html to dataframe with text}
\usage{
from_html_2_df(html_page)
}
\arguments{
\item{html_page}{the pure html text of the tacquigraphic notes. It can be the
html text as well as the "xml_document". It does not accept the URL as input

file_html <- list.files("inst/extdata/", full.names=TRUE)
NT_html <- xml2::read_html(file_html,encoding = "utf8")
from_html_2_df(NT_html)}
}
\description{
turns the pure html text of the tacquigraphic notes, into
an structured tibble/ data.frame with two columns: title and text. Te results
from this function is send further to the parser function
}
