% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{func_DB_NT}
\alias{func_DB_NT}
\title{Parser de notas taquigráficas do Senado}
\usage{
func_DB_NT(linha)
}
\description{
Função que faz o parser da nota taquigráfica, o estruturando em um data.frame salvo em rds e csv
}
