% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tn.R
\name{get_all_tn}
\alias{get_all_tn}
\title{get_all_tn}
\usage{
get_all_tn(DF)
}
\arguments{
\item{DF}{The dataframe generated by reunion()}
}
\description{
get all  tacquigraphic notes from one commission (obtained by reunion() ) and
structures all in one dataframe
}
\examples{
get_all_NT(DF)

}
