% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{get_info}
\usage{
get_info(cod)
}
\arguments{
\item{cod}{the number of the code of the commission}
}
\description{
returns a list with some info about the commission, like title, description
date of installation (start_date)
}
\examples{
get_info(2292)

}
