% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parties.R
\name{get_parties}
\alias{get_parties}
\title{Get Brazilian political parties in the Senate History}
\usage{
get_parties()
}
\value{
tibble
}
\description{
get a dataframe with existing and extinguished Brazilian political parties in Senate.
"Obtém a lista dos Partidos Políticos em atividade e/ou extintos no Senado Federal"
}
\examples{
get_parties()
}
