% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse the tacquigraphic notes text}
\usage{
parse(html_page, save_as)
}
\arguments{
\item{html_page}{the dataframe generated with meetings()}
}
\description{
turns the pure .html text of the tacquigraphic notes, into an structured
tibble/ data.frame.
}
\examples{
df <- meetings(cod = "2606", start = "2023-05-25")
parser( df, 1 )

}
