% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_TN}
\alias{parse_TN}
\title{Parse the tacquigraphic notes text}
\usage{
parse_TN(html_page, save_as)
}
\arguments{
\item{html_page}{the dataframe generated with meetings()}

\item{save_as}{To save the files in .rds and .csv. If empty, the files will not be saved.}
}
\description{
turns the pure .html text of the tacquigraphic notes, into an structured
tibble/ data.frame.
}
\examples{
# parse_NT("https://www25.senado.leg.br/web/atividade/notas-taquigraficas/-/notas/r/13611")
file_html <- list.files("inst/extdata/", full.names=TRUE)
TN_html <- xml2::read_html(file_html,encoding = "utf8")
parse_TN(TN_html)
}
