% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser}
\alias{parser}
\title{Parse the tacquigraphic notes text}
\usage{
parser(DF, linha, save = FALSE)
}
\arguments{
\item{DF}{the dataframe generated with meetings()}

\item{linha}{number of the line of the dataframe generated with meetings()}

\item{save}{= FALSE (default). To save the files in .rds and .csv.}
}
\description{
turns the pure html text of the tacquigraphic notes, into
an structured tibble/ data.frame, and saves it in .rds and .csv
}
\examples{
df <- meetings(cod = "2606", start = "2023-05-25")
parser( df, 1 )

}
