% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{save_csv_rds}
\alias{save_csv_rds}
\title{Save structured data from extract_metadata}
\usage{
save_csv_rds(NTDB, save_as = NULL)
}
\arguments{
\item{save_as}{= "rds" (default). To save the files in .rds and .csv.}

\item{DF}{the dataframe generated by extract_metadata}
}
\description{
input is DF generated by from_html_2_df(), export a DF with more columns
}
