% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\docType{data}
\name{TodosPartidos}
\alias{TodosPartidos}
\title{Parser de notas taquigráficas do Senado}
\format{
An object of class \code{character} of length 43.
}
\usage{
TodosPartidos
}
\description{
All brazilian parties
}
\details{
The names of parties used to extract the data of the column "partido"
Estes nomes serão usados para construir a coluna "partido"
}
\examples{
TodosPartidos

}
\keyword{datasets}
