% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse}
\alias{parse}
\title{input is DF generated by from_html_2_df()}
\usage{
parse(DF)
}
\description{
input is DF generated by from_html_2_df()
}
\examples{
file_html <- list.files("inst/extdata/", full.names=TRUE)
NT_html <- xml2::read_html(file_html,encoding = "utf8")
parse(NT_html)
}
