% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reunioes.R
\name{reunions}
\alias{reunions}
\title{lendo a página html com todas as reuniões e retorna um tibble}
\usage{
reunions(cod, start, end = Sys.Date())
}
\arguments{
\item{cod}{The code of the reunion}

\item{start}{The start date}

\item{end}{The end date, If no parameter is given, it uses the actual date}
}
\description{
lendo a página html com todas as reuniões e retorna um tibble
}
\examples{
reunions(cod = "2606", start = "2023-05-25", end ="2023-08-12")

}
